reorganize keyring tests to match #100 and keyring namespace to simplify names
