+++ src/aws_encryption_sdk/keyrings/base.py
