+++ src/aws_encryption_sdk/keyrings/raw.py
